backend/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
SQLAlchemy
python-dotenv
openai
bcrypt
